+++ src/include/postmaster/pagewriter.h
@@ -104,7 +104,7 @@ extern void ckpt_shutdown_pagewriter();
 extern uint64 get_dirty_page_queue_rec_lsn();
 extern XLogRecPtr ckpt_get_min_rec_lsn(void);
 extern uint32 calculate_thread_max_flush_num(bool is_pagewriter);
-extern uint32 get_loc_for_lsn(XLogRecPtr target_lsn);
+extern uint64 get_loc_for_lsn(XLogRecPtr target_lsn);
 extern uint64 get_time_ms();
 
 const int PAGEWRITER_VIEW_COL_NUM = 8;
